backend/requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
numpy==1.21.2
scipy==1.7.1
Pillow==8.3.2
werkzeug==2.0.1
python-dotenv==0.19.0
flask-cors==3.0.10
dlib==19.22.0
face-recognition==1.3.0 